cs490/requirements.txt
cffi==1.9.1
cryptography==1.5.3
enum34==1.1.6
futures==3.0.5
idna==2.1
ipaddress==1.0.17
pyasn1==0.1.9
pycparser==2.17
pyOpenSSL==16.2.0
requests==2.12.0
six==1.10.0
